moove all age logic into a dedicated trait
